requirements.txt
Django==1.7.2
Pillow==2.7.0
argparse==1.2.1
django-admin-tools==0.5.2
django-ckeditor==4.4.7
django-contrib-comments==1.5
django-filebrowser==3.5.7
django-grappelli==2.5.6
django-mptt==0.6.1
django-tagging==0.3.4
django-wpadmin==1.7.0
django-xmlrpc==0.1.5
oauthlib==0.7.2
pyparsing==2.0.3
python-openid==2.2.5
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
wsgiref==0.1.2
